WIP/PoC: Enable third-parties to provide variables in config files
Make it possible for third-parties to provide a file containing
prefined substitution variable for debhelper config files.  This is
done by installing the files as:

    /usr/share/debhelper/extension-snippets/<package>.config-subst

In this demo, debhelper also provides such a variable enabling the
path to be defined as:

    ${ext:debhelper:datafiles_dir}/extension-snippets/<package>.config-subst

These files consist of key-value lines of the form:

    <package>:<variable>=<value>

The consumer can then refer to them as `${ext:<package>:<variable>}`
and get `<value>` as replacement.  Note that `<value>` can contain
substitution variables and these are replaced as well (at most one
level deep - recursion is not supported).

Signed-off-by: Niels Thykier <niels@thykier.net>