Dh_Buildsystems: Add bmake and mkcmake as third-party build systems
mk-configure and bmake ship autodetection functions which can tell
when one of them is being used based on specific lines in the Makefile
which GNU Make doesn’t support. More specifically:

.include <…> or include <…> (with angle brackets)

    is for bmake

.include <bsd.…> or include <bsd.…>

    is for bmake with NetBSD mk files (pmake mode)

.include <mkc…>

    is for mk-configure

Adding bmake and mkcmake to the third-party list makes it possible
to remove the explicit --buildsystem argument to dh for most cases
except those when makefiles are written specifically in bmake syntax,
but don’t use any of the bmake, NetBSD or mk-configure mk files.

Signed-off-by: Andrej Shadura <andrew.shadura@collabora.co.uk>